Updated test.ref.m2
In the previous m2 file, annotator disagreements are not captured correctly. If annotators disagree on the grammaticality of the sentence, a '-noop-' edit should be added for the annotator who annotated the sentence as grammatical. This is important to capture as "unnecessary edits " from a single annotator could impact the evaluation.

This approach of adding "-noop" edit is followed in conll-14 & other datasets.